requirements.txt
# LLM
openai>=1.14.0

# LangChain core
langchain==0.3.27
langchain-core>=0.3.72,<0.4.0
langchain-community>=0.3.30,<0.4.0
langchain-text-splitters>=0.3.11,<0.4.0

# LangChain integrations
langchain-openai>=0.3.33,<0.4.0
langchain-chroma==0.2.6

# Vector DB (MUST be 1.x)
chromadb>=1.0.20,<2.0.0

# UI
gradio>=5.47.2

# Utilities
python-dotenv>=1.0.1
pydantic>=2.7.4
numpy>=1.26.4
tenacity>=8.3.0
